move test setup code into an Internals module
