+++ test/batch.js
@@ -3,17 +3,10 @@
 // Load modules
 
 const Async = require('async');
-const Bassmaster = require('../');
 const Code = require('code');
-const Hapi = require('hapi');
 const Lab = require('lab');
 const Sinon = require('sinon');
-
-
-// Declare internals
-
-const internals = {};
-
+const Internals = require('./internals.js');
 
 // Test shortcuts
 
@@ -23,197 +16,18 @@ const it = lab.it;
 const before = lab.before;
 const expect = Code.expect;
 
+let server = null;
 
 describe('Batch', () => {
 
-    let server = null;
+    before((done) => {
 
-    const profileHandler = function (request, reply) {
-
-        const id = request.query.id || 'fa0dbda9b1b';
-
-        return reply({
-            'id': id,
-            'name': 'John Doe'
-        });
-    };
-
-    const activeItemHandler = function (request, reply) {
-
-        return reply({
-            'id': '55cf687663',
-            'name': 'Active Item'
-        });
-    };
-
-    const itemHandler = function (request, reply) {
-
-        return reply({
-            'id': request.params.id,
-            'name': 'Item'
-        });
-    };
-
-    const item2Handler = function (request, reply) {
-
-        return reply({
-            'id': request.params.id || 'mystery-guest',
-            'name': 'Item'
-        });
-    };
-
-    const arrayHandler = function (request, reply) {
-
-        return reply({
-            'id': '55cf687663',
-            'name': 'Dress',
-            'items': [{ 'color': 'blackandblue' }, { 'color': 'whiteandgold' }]
-        });
-    };
-
-    const zeroIntegerHandler = function (request, reply) {
-
-        return reply({
-            'id': 0,
-            'name': 'Zero Item'
-        });
-    };
-
-    const integerHandler = function (request, reply) {
-
-        return reply({
-            'id': 123,
-            'name': 'Integer Item'
-        });
-    };
-
-    const integerItemHandler = function (request, reply) {
-
-        return reply({
-            'id': request.params.id,
-            'name': 'Integer'
-        });
-    };
-
-    const badCharHandler = function (request, reply) {
-
-        return reply({
-            'id': 'test',
-            'null': null,
-            'invalidChar': '#'
-        });
-    };
-
-    const badValueHandler = function (request, reply) {
-
-        return reply(null);
-    };
-
-    const redirectHandler = function (request, reply) {
-
-        return reply().redirect('/profile');
-    };
-
-    const fetch1 = function (request, next) {
-
-        next('Hello');
-    };
-
-    const fetch2 = function (request, next) {
-
-        next(request.pre.m1 + request.pre.m3 + request.pre.m4);
-    };
-
-    const fetch3 = function (request, next) {
-
-        process.nextTick(() => {
-
-            next(' ');
-        });
-    };
-
-    const fetch4 = function (request, next) {
-
-        next('World');
-    };
-
-    const fetch5 = function (request, next) {
-
-        next(request.pre.m2 + '!');
-    };
-
-    const getFetch = function (request, reply) {
-
-        return reply(request.pre.m5 + '\n');
-    };
-
-    const errorHandler = function (request, reply) {
-
-        return reply(new Error('myerror'));
-    };
-
-    const echoHandler = function (request, reply) {
-
-        return reply(request.payload);
-    };
-
-    const setupServer = function (done) {
-
-        server = new Hapi.Server();
-        server.connection();
-        server.route([
-            { method: 'POST', path: '/echo', handler: echoHandler },
-            { method: 'PUT', path: '/echo', handler: echoHandler },
-            { method: 'GET', path: '/profile', handler: profileHandler },
-            { method: 'GET', path: '/item', handler: activeItemHandler },
-            { method: 'GET', path: '/array', handler: arrayHandler },
-            { method: 'GET', path: '/item/{id}', handler: itemHandler },
-            { method: 'GET', path: '/item2/{id?}', handler: item2Handler },
-            { method: 'GET', path: '/zero', handler: zeroIntegerHandler },
-            { method: 'GET', path: '/int', handler: integerHandler },
-            { method: 'GET', path: '/int/{id}', handler: integerItemHandler },
-            { method: 'GET', path: '/error', handler: errorHandler },
-            { method: 'GET', path: '/badchar', handler: badCharHandler },
-            { method: 'GET', path: '/badvalue', handler: badValueHandler },
-            {
-                method: 'GET',
-                path: '/fetch',
-                handler: getFetch,
-                config: {
-                    pre: [
-                        { method: fetch1, assign: 'm1', mode: 'parallel' },
-                        { method: fetch2, assign: 'm2' },
-                        { method: fetch3, assign: 'm3', mode: 'parallel' },
-                        { method: fetch4, assign: 'm4', mode: 'parallel' },
-                        { method: fetch5, assign: 'm5' }
-                    ]
-                }
-            },
-            { method: 'GET', path: '/redirect', handler: redirectHandler }
-        ]);
-
-        server.register(Bassmaster, done);
-    };
-
-    const makeRequest = function (payload, callback) {
-
-        const next = function (res) {
-
-            return callback(res.result);
-        };
-
-        server.connections[0].inject({
-            method: 'post',
-            url: '/batch',
-            payload: payload
-        }, next);
-    };
-
-    before(setupServer);
+        server = Internals.setupServer(done);
+    });
 
     it('shows single response when making request for single endpoint', (done) => {
 
-        makeRequest('{ "requests": [{ "method": "get", "path": "/profile" }] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{ "method": "get", "path": "/profile" }] }', (res) => {
 
             expect(res[0].id).to.equal('fa0dbda9b1b');
             expect(res[0].name).to.equal('John Doe');
@@ -224,7 +38,7 @@ describe('Batch', () => {
 
     it('supports redirect', (done) => {
 
-        makeRequest('{ "requests": [{ "method": "get", "path": "/redirect" }] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{ "method": "get", "path": "/redirect" }] }', (res) => {
 
             expect(res[0].id).to.equal('fa0dbda9b1b');
             expect(res[0].name).to.equal('John Doe');
@@ -235,7 +49,7 @@ describe('Batch', () => {
 
     it('supports query string in the request', (done) => {
 
-        makeRequest('{ "requests": [{ "method": "get", "path": "/profile?id=someid" }] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{ "method": "get", "path": "/profile?id=someid" }] }', (res) => {
 
             expect(res[0].id).to.equal('someid');
             expect(res[0].name).to.equal('John Doe');
@@ -246,7 +60,7 @@ describe('Batch', () => {
 
     it('supports non alphanum characters in the request', (done) => {
 
-        makeRequest('{ "requests": [{ "method": "get", "path": "/item/item-_^~&-end" }] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{ "method": "get", "path": "/item/item-_^~&-end" }] }', (res) => {
 
             expect(res[0].id).to.equal('item-_^~&-end');
             expect(res[0].name).to.equal('Item');
@@ -257,7 +71,7 @@ describe('Batch', () => {
 
     it('shows two ordered responses when requesting two endpoints', (done) => {
 
-        makeRequest('{ "requests": [{"method": "get", "path": "/profile"}, {"method": "get", "path": "/item"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{"method": "get", "path": "/profile"}, {"method": "get", "path": "/item"}] }', (res) => {
 
             expect(res[0].id).to.equal('fa0dbda9b1b');
             expect(res[0].name).to.equal('John Doe');
@@ -270,7 +84,7 @@ describe('Batch', () => {
 
     it('shows two ordered responses when requesting two endpoints (with optional path param)', (done) => {
 
-        makeRequest('{ "requests": [{"method": "get", "path": "/item2/john"}, {"method": "get", "path": "/item2/"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{"method": "get", "path": "/item2/john"}, {"method": "get", "path": "/item2/"}] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].id).to.equal('john');
@@ -364,7 +178,7 @@ describe('Batch', () => {
             '] }';
 
         const asyncSpy = Sinon.spy(Async, 'parallel');
-        makeRequest(requestBody, (res) => {
+        Internals.makeRequest(server, requestBody, (res) => {
 
             expect(res[0].id).to.equal('fa0dbda9b1b');
             expect(res[0].name).to.equal('John Doe');
@@ -378,7 +192,7 @@ describe('Batch', () => {
 
     it('supports piping a response into the next request', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "get", "path": "/item"}, {"method": "get", "path": "/item/$0.id"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "get", "path": "/item"}, {"method": "get", "path": "/item/$0.id"}] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].id).to.equal('55cf687663');
@@ -391,7 +205,7 @@ describe('Batch', () => {
 
     it('supports piping integer response into the next request', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "get", "path": "/int"}, {"method": "get", "path": "/int/$0.id"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "get", "path": "/int"}, {"method": "get", "path": "/int/$0.id"}] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].id).to.equal(123);
@@ -404,7 +218,7 @@ describe('Batch', () => {
 
     it('supports piping a zero integer response into the next request', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "get", "path": "/zero"}, {"method": "get", "path": "/int/$0.id"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "get", "path": "/zero"}, {"method": "get", "path": "/int/$0.id"}] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].id).to.equal(0);
@@ -417,7 +231,7 @@ describe('Batch', () => {
 
     it('supports posting multiple requests', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "post", "path": "/echo", "payload":{"a":1}}, {"method": "post", "path": "/echo", "payload":{"a":2}}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "post", "path": "/echo", "payload":{"a":1}}, {"method": "post", "path": "/echo", "payload":{"a":2}}] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0]).to.deep.equal({ a: 1 });
@@ -428,7 +242,7 @@ describe('Batch', () => {
 
     it('supports sending multiple PUTs requests', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "put", "path": "/echo", "payload":{"a":1}}, {"method": "put", "path": "/echo", "payload":{"a":2}}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "put", "path": "/echo", "payload":{"a":1}}, {"method": "put", "path": "/echo", "payload":{"a":2}}] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0]).to.deep.equal({ a: 1 });
@@ -439,7 +253,7 @@ describe('Batch', () => {
 
     it('supports piping a response from post into the next get request', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "post", "path": "/echo", "payload": {"id":"55cf687663"}}, {"method": "get", "path": "/item/$0.id"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "post", "path": "/echo", "payload": {"id":"55cf687663"}}, {"method": "get", "path": "/item/$0.id"}] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].id).to.equal('55cf687663');
@@ -451,7 +265,7 @@ describe('Batch', () => {
 
     it('supports piping a nested response value from post into the next get request', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "post", "path": "/echo", "payload": { "data": {"id":"44cf687663"}}}, {"method": "get", "path": "/item/$0.data.id"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "post", "path": "/echo", "payload": { "data": {"id":"44cf687663"}}}, {"method": "get", "path": "/item/$0.data.id"}] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].data.id).to.equal('44cf687663');
@@ -463,7 +277,7 @@ describe('Batch', () => {
 
     it('handles null payloads gracefully', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "post", "path": "/echo", "payload":{"a":1}}, {"method": "post", "path": "/echo", "payload":null}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "post", "path": "/echo", "payload":{"a":1}}, {"method": "post", "path": "/echo", "payload":null}] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0]).to.deep.equal({ a: 1 });
@@ -474,7 +288,7 @@ describe('Batch', () => {
 
     it('includes errors when they occur in the request', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "get", "path": "/item"}, {"method": "get", "path": "/nothere"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "get", "path": "/item"}, {"method": "get", "path": "/nothere"}] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].id).to.equal('55cf687663');
@@ -486,7 +300,7 @@ describe('Batch', () => {
 
     it('bad requests return the correct error', (done) => {
 
-        makeRequest('{ "blah": "test" }', (res) => {
+        Internals.makeRequest(server, '{ "blah": "test" }', (res) => {
 
             expect(res.statusCode).to.equal(400);
             done();
@@ -496,7 +310,7 @@ describe('Batch', () => {
 
     it('handles empty payload', (done) => {
 
-        makeRequest(null, (res) => {
+        Internals.makeRequest(server, null, (res) => {
 
             expect(res.statusCode).to.equal(400);
             done();
@@ -505,7 +319,7 @@ describe('Batch', () => {
 
     it('handles payload request not array', (done) => {
 
-        makeRequest('{ "requests": {"method": "get", "path": "/$1"} }', (res) => {
+        Internals.makeRequest(server, '{ "requests": {"method": "get", "path": "/$1"} }', (res) => {
 
             expect(res.statusCode).to.equal(400);
             done();
@@ -514,7 +328,7 @@ describe('Batch', () => {
 
     it('handles bad paths in requests array', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "get", "path": "/$1"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "get", "path": "/$1"}] }', (res) => {
 
             expect(res.statusCode).to.equal(400);
             done();
@@ -523,7 +337,7 @@ describe('Batch', () => {
 
     it('handles errors in the requested handlers', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "get", "path": "/error"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "get", "path": "/error"}] }', (res) => {
 
             expect(res[0].statusCode).to.equal(500);
             done();
@@ -532,7 +346,7 @@ describe('Batch', () => {
 
     it('an out of bounds reference returns an error', (done) => {
 
-        makeRequest('{ "requests": [{"method": "get", "path": "/item"}, {"method": "get", "path": "/item/$1.id"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{"method": "get", "path": "/item"}, {"method": "get", "path": "/item/$1.id"}] }', (res) => {
 
             expect(res.error).to.equal('Bad Request');
             done();
@@ -541,7 +355,7 @@ describe('Batch', () => {
 
     it('a non-existant reference returns an internal error', (done) => {
 
-        makeRequest('{ "requests": [{"method": "get", "path": "/item"}, {"method": "get", "path": "/item/$0.nothere"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{"method": "get", "path": "/item"}, {"method": "get", "path": "/item/$0.nothere"}] }', (res) => {
 
             expect(res.statusCode).to.equal(500);
             done();
@@ -550,7 +364,7 @@ describe('Batch', () => {
 
     it('a non-existant & nested reference returns an internal error', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "post", "path": "/echo", "payload": { "data": {"id":"44cf687663"}}}, {"method": "get", "path": "/item/$0.data.not.here"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "post", "path": "/echo", "payload": { "data": {"id":"44cf687663"}}}, {"method": "get", "path": "/item/$0.data.not.here"}] }', (res) => {
 
             expect(res.statusCode).to.equal(500);
             done();
@@ -559,7 +373,7 @@ describe('Batch', () => {
 
     it('handles a bad character in the reference value', (done) => {
 
-        makeRequest('{ "requests": [{"method": "get", "path": "/badchar"}, {"method": "get", "path": "/item/$0.invalidChar"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{"method": "get", "path": "/badchar"}, {"method": "get", "path": "/item/$0.invalidChar"}] }', (res) => {
 
             expect(res.statusCode).to.equal(500);
             done();
@@ -568,7 +382,7 @@ describe('Batch', () => {
 
     it('handles a null value in the reference value', (done) => {
 
-        makeRequest('{ "requests": [{"method": "get", "path": "/badchar"}, {"method": "get", "path": "/item/$0.null"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{"method": "get", "path": "/badchar"}, {"method": "get", "path": "/item/$0.null"}] }', (res) => {
 
             expect(res.statusCode).to.equal(500);
             done();
@@ -577,7 +391,7 @@ describe('Batch', () => {
 
     it('cannot use invalid character to request reference', (done) => {
 
-        makeRequest('{ "requests": [{"method": "get", "path": "/badvalue"}, {"method": "get", "path": "/item/$:.name"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{"method": "get", "path": "/badvalue"}, {"method": "get", "path": "/item/$:.name"}] }', (res) => {
 
             expect(res.statusCode).to.equal(400);
             done();
@@ -586,7 +400,7 @@ describe('Batch', () => {
 
     it('handles missing reference', (done) => {
 
-        makeRequest('{ "requests": [{"method": "get", "path": "/badvalue"}, {"method": "get", "path": "/item/$0.name"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{"method": "get", "path": "/badvalue"}, {"method": "get", "path": "/item/$0.name"}] }', (res) => {
 
             expect(res.statusCode).to.equal(500);
             done();
@@ -595,7 +409,7 @@ describe('Batch', () => {
 
     it('handles error when getting reference value', (done) => {
 
-        makeRequest('{ "requests": [{"method": "get", "path": "/item"}, {"method": "get", "path": "/item/$0.1"}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{"method": "get", "path": "/item"}, {"method": "get", "path": "/item/$0.1"}] }', (res) => {
 
             expect(res.statusCode).to.equal(500);
             done();
@@ -604,7 +418,7 @@ describe('Batch', () => {
 
     it('supports an optional query object', (done) => {
 
-        makeRequest('{ "requests": [{ "method": "get", "path": "/profile", "query": { "id": "someid" } }] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{ "method": "get", "path": "/profile", "query": { "id": "someid" } }] }', (res) => {
 
             expect(res[0].id).to.equal('someid');
             expect(res[0].name).to.equal('John Doe');
@@ -615,7 +429,7 @@ describe('Batch', () => {
 
     it('supports alphanum characters in the query', (done) => {
 
-        makeRequest('{ "requests": [{ "method": "get", "path": "/profile", "query": { "id": "item-_^~&-end" } }] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [{ "method": "get", "path": "/profile", "query": { "id": "item-_^~&-end" } }] }', (res) => {
 
             expect(res[0].id).to.equal('item-_^~&-end');
             expect(res[0].name).to.equal('John Doe');
@@ -626,7 +440,7 @@ describe('Batch', () => {
 
     it('handles null queries gracefully', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "post", "path": "/echo", "query": null}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "post", "path": "/echo", "query": null}] }', (res) => {
 
             expect(res.length).to.equal(1);
             expect(res[0]).to.deep.equal(null);
@@ -636,7 +450,7 @@ describe('Batch', () => {
 
     it('supports piping a whole payload to the next request', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "get", "path": "/item"}, {"method": "post", "path": "/echo", "payload":"$0"} ] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "get", "path": "/item"}, {"method": "post", "path": "/echo", "payload":"$0"} ] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].id).to.equal('55cf687663');
@@ -649,7 +463,7 @@ describe('Batch', () => {
 
     it('supports piping a partial payload to the next request', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "get", "path": "/item"}, {"method": "post", "path": "/echo", "payload":"$0.name"} ] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "get", "path": "/item"}, {"method": "post", "path": "/echo", "payload":"$0.name"} ] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].id).to.equal('55cf687663');
@@ -661,7 +475,7 @@ describe('Batch', () => {
 
     it('supports piping a partial payload from a nested array to the next request', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "get", "path": "/array"}, {"method": "post", "path": "/echo", "payload":"$0.items.1"} ] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "get", "path": "/array"}, {"method": "post", "path": "/echo", "payload":"$0.items.1"} ] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].id).to.equal('55cf687663');
@@ -673,7 +487,7 @@ describe('Batch', () => {
 
     it('returns an empty object when a non-existent path is set at the root of the payload', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "get", "path": "/item"}, {"method": "post", "path": "/echo", "payload":"$0.foo"} ] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "get", "path": "/item"}, {"method": "post", "path": "/echo", "payload":"$0.foo"} ] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].id).to.equal('55cf687663');
@@ -685,7 +499,7 @@ describe('Batch', () => {
 
     it('sets a nested reference in the payload', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "get", "path": "/item"}, {"method": "post", "path": "/echo", "payload":{"name2": "$0.name"}} ] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "get", "path": "/item"}, {"method": "post", "path": "/echo", "payload":{"name2": "$0.name"}} ] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].id).to.equal('55cf687663');
@@ -697,7 +511,7 @@ describe('Batch', () => {
 
     it('returns an empty object when a nonexistent path is set in the payload', (done) => {
 
-        makeRequest('{ "requests": [ {"method": "get", "path": "/item"}, {"method": "post", "path": "/echo", "payload":{"foo": "$0.foo"}} ] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "get", "path": "/item"}, {"method": "post", "path": "/echo", "payload":{"foo": "$0.foo"}} ] }', (res) => {
 
             expect(res.length).to.equal(2);
             expect(res[0].id).to.equal('55cf687663');
@@ -713,7 +527,7 @@ describe('Batch', () => {
         // Add a connection to the server
         server.connection({ port: 8000, host: 'localhost', labels: ['test'] });
 
-        makeRequest('{ "requests": [ {"method": "post", "path": "/echo", "query": null}] }', (res) => {
+        Internals.makeRequest(server, '{ "requests": [ {"method": "post", "path": "/echo", "query": null}] }', (res) => {
 
             expect(res.length).to.equal(1);
             expect(res[0]).to.deep.equal(null);
